feat: Add authentication and refresh token to the fastify-jwt module
The FastifyJWT module now has two additional properties: hashing for the user and payload objects, and createdAt for timestamp of JWT creation. The modules IConfig class has also been updated to contain access and refresh token time-to-live values. Additionally, session is renamed to auth and an endpoint to create a refresh token and an endpoint to refresh an access token has been introduced. Lastly, a function to generate a new refresh token has been added to the Utils class.