Refactor auth and refresh token methods for clearer token creation and management
This commit refactors the `/auth.init` and `/session.getRefreshToken` API endpoints to use a more standardized and streamlined approach to creating access and refresh tokens through the use of the `createRefreshToken` method in `utils.ts`. In addition, a new `accessToken` field has been added to the `refreshTokenSchema` schema to track which access token is associated with which refresh token.