Refactor security API methods and incident schema
This commit refactors the security API methods allowAccessToArea and denyAccessToArea to add a passLogId field to the addSecurityIncident method. Also, it removes the redundant security checking to distinguish between granting and denying access. Finally, it includes a modified securityIncident schema with a new passLogId field for additional tracking of user's pass log.