feat(devices): Add support for IP address in device creation API endpoint
This commit adds support for two additional parameters 'ip' and 'isEnabled' to the device creation API endpoint. An IP address validation check has been added to ensure that it passes the format required for an IP address. The 'isEnabled' field is now optional and can be set to true/false.